Airport Fully Tested

+++ Runway_Redeclaration/src/Core/DeclaredRunway.java
@@ -110,7 +110,7 @@ import Exceptions.VariableDeclarationException;
 		if(angle<100){
 			out += "0";
 		}
-		out += String.valueOf(Math.round(angle/10))+direction;
+		out += String.valueOf(Math.round((double)angle/10))+direction;
 		return out;
 	}
 //----[ Setters ]---------------------------------------------------s---
